final run for real test
We decide to use the model obtained at epoch 15. We run 50 epochs to see and get insights on learning curves on test set, but we will not take any decision from that. Decision to consider epoch 15 is derived by the investigation of the previous CV runs (files `runs/<yyyymmddHHMMSS>_run.rds`).